requirements.txt
io
requests
pytesseract
beautifulsoup4
Pillow
aiohttp
